Create the basic database
Script developed on MySQL to create a DB called "Salinha", which contains three tables. A table called "Produto", to store data about the products. Another table called "Estoque", to store quantity and the last time it was updated. And then, a table for the update's historic, containing what product was updated, how much was update, if the update was an input or output and the responsible for that update.